ap-dnsproxy: move iptables command to puavo-wlanap-dnsproxy
Previously, when iptables was modified directly from vtun-up/down
scripts, iptables was modified even when puavo-wlanap-dnsproxy was not
running. Now scripts just notify the main daemon about vtun-up/down
events, and the daemon is responsible of applying necessary iptables
changes. If puavo-wlanap-dnsproxy is not running, nothing happens, as it
should be.