mapper: use horst instead of tshark for RF-monitoring
Horst is really lightweight and fast (tsharks's startup time is really bad) and
has implemented channel hopping so that we do not have to.

Horst can also be run in "server" mode and controlled with simple commands via a
control pipe. Now, we are just starting/stopping horst whenever we want to
measure, but we could just pause/resume the single process and optionally save
output of those measurement sessions to spot-specific output files.